Update AI.java
Some method name changes, but mainly this commit involved the creation of the ParPos class that determines and stores the positions of parentheses in an expression, and also determines and stores if there is an error in the expression. This is used in several methods and reduced the number of lines of code by about 70.